🧹 move new tests into __tests__ per convention
PR #312 co-located test files next to source instead of the repo's
__tests__ convention; fix relative imports accordingly.
