+++ libs/backend/agents/personal/session/main/src/__tests__/fleet-membership-identity-envelope-source.test.ts
@@ -1,9 +1,9 @@
 import type { FleetMembershipSignatureVerifier } from "@opencrane/backend/server/iam/membership";
 import type { InitialRunAuthority, RunAdmissionTransaction } from "@opencrane/backend/agents/personal/runs";
-import type { CapabilitySetDigestSource, SessionAssemblyCommand } from "./session-assembly.types.js";
+import type { CapabilitySetDigestSource, SessionAssemblyCommand } from "../session-assembly.types.js";
 import { describe, expect, it, vi } from "vitest";
 
-import { FleetMembershipIdentityEnvelopeSource } from "./fleet-membership-identity-envelope-source.js";
+import { FleetMembershipIdentityEnvelopeSource } from "../fleet-membership-identity-envelope-source.js";
 
 /** Creates one final-admission command bound to the exact signed assertion fixture. */
 function _command(): SessionAssemblyCommand
